servers/Flask/requirements.txt
flask==3.1.2
langchain==0.2.6
youtube_transcript_api==1.2.1
ibm_watsonx_ai== 1.0.10
langchain_ibm==0.1.8
langchain-community==0.2.6
python-dotenv==1.2.1
flask_cors==6.0.2